pyproject.toml
[project]
name = "hackathon-huggingface"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "smolagents[litellm]>=1.18.0",
    "streamlit>=1.45.1",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest-dotenv>=0.5.2",
    "pytest>=8.3.3",
    "ruff>=0.7.4",
]

[tool.setuptools]
packages = ["workspace"]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_untyped_decorators = true

files = ["workspace/src"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
env_files = [".env"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
pythonpath = [
  "workspace/src",
]
